Refactored snippet filter label code
