+++ app/client/src/components/editorComponents/GlobalSearch/SnippetRefinements.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { connectCurrentRefinements } from "react-instantsearch-dom";
 import styled from "styled-components";
 import { ReactComponent as CloseIcon } from "assets/icons/help/close_blue.svg";
-import { FilterEntity, SnippetFilterLabel } from "./SnippetsFilter";
+import { getSnippetFilterLabel } from "./utils";
 
 const RefinementListContainer = styled.div`
   background: ${(props) => props.theme.colors.globalSearch.primaryBgColor};
@@ -42,14 +42,7 @@ const RefinementListContainer = styled.div`
 function RefinementPill({ item, refine }: any) {
   return (
     <div className="refinement-pill">
-      <span>
-        {SnippetFilterLabel[item.label as FilterEntity] ||
-          item.label
-            .toLowerCase()
-            .replace("_widget", "")
-            .replace("-plugin", "")
-            .replaceAll(/_|-/g, " ")}
-      </span>
+      <span>{getSnippetFilterLabel(item.label)}</span>
       <CloseIcon
         onClick={(event) => {
           event.preventDefault();
